src/localguide/pyproject.toml
[project]
name = "localguide"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic-ai-slim[openai]>=0.2.3",
    "uvicorn>=0.34.2",
]
